+++ game_patch/graphics/gr.cpp
@@ -123,6 +123,18 @@ ConsoleCommand2 fov_cmd{
     "fov [degrees]",
 };
 
+ConsoleCommand2 gamma_cmd{
+    "gamma",
+    [](std::optional<float> value_opt) {
+        if (value_opt) {
+            rf::gr::set_gamma(value_opt.value());
+        }
+        rf::console::printf("Gamma: %.2f", rf::gr::gamma);
+    },
+    "Sets gamma.",
+    "gamma [value]",
+};
+
 CallHook<void(int, rf::gr::Vertex**, int, rf::gr::Mode)> gr_rect_gr_tmapper_hook{
     0x0050DD69,
     [](int nv, rf::gr::Vertex** verts, int flags, rf::gr::Mode mode) {
@@ -266,4 +278,5 @@ void gr_apply_patch()
 
     // Commands
     fov_cmd.register_cmd();
+    gamma_cmd.register_cmd();
 }